+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Fixed
+- `field-create`: Add clear error message if there are no existing fields to be added
+
 ## [1.12.0] - 2021-09-02
 ### Added
 - Add model method generators for field types of the `computed_field` module.